+++ src/components/Main/Selector/SelectorForm/SelectorForm.jsx
@@ -6,10 +6,13 @@ import {
   IoArrowForwardCircleOutline,
   IoCheckmarkCircleOutline,
 } from "react-icons/io5";
+import { CiClock1, CiClock2, CiTimer } from "react-icons/ci";
 import Button from "./Button/Button.jsx";
 import Label from "./Label/Label.jsx";
 import SelectKey from "./Select/SelectKey.jsx";
 import SelectValue from "./Select/SelectValue.jsx";
+import SelectInterval from "./Select/SelectInterval.jsx";
+import DatePicker from "./DatePicker/DatePicker.jsx";
 import { apiCallToGetCalculatedMetrics } from "../../../../api/Api.jsx";
 import $SelectorForm, { $ErrorLabel } from "./SelectorForm.styles.jsx";
 
@@ -25,6 +28,8 @@ export default function SelectorForm() {
     whereCriteria,
     setWhereCriteria,
     setCalculatedMetrics,
+    setPeriod,
+    period
   } = useContext(SelectorContext);
 
   function handleShowCriteriaChange(event) {
@@ -33,6 +38,10 @@ export default function SelectorForm() {
 
   function handleWhereCriteriaChange(event) {
     setTempWhereCriteria({ key: event.target.value, value: "", id: "" });
+  }
+
+  function handlePeriodIntervalChange(event) {
+    setPeriod({interval: event.target.value});
   }
 
   async function fetchCalculatedMetrics(event) {
@@ -45,15 +54,15 @@ export default function SelectorForm() {
 
     const moment = require("moment");
     const date = new Date();
-    const startDate = moment().startOf("month").format("DD-MM-YYYY HH:mm:ss");
-    const endDate = moment(date).format("DD-MM-YYYY HH:mm:ss");
-    const period = 900;
+    const startDate = period.start !== null ? period.start : moment().startOf("month").format("DD-MM-YYYY HH:mm:ss");
+    const endDate = period.end !== null ? period.end : moment(date).format("DD-MM-YYYY HH:mm:ss");
+    const interval = period.interval !== null ? period.interval : 900;
 
     await apiCallToGetCalculatedMetrics(
       showCriteria,
       startDate,
       endDate,
-      period,
+      interval,
       whereCriteriaMap
     )
       .then((resData) => {
@@ -97,6 +106,19 @@ export default function SelectorForm() {
         >
           <IoAddCircleOutline />
         </Button>
+        <div style={{ flexBasis: '100%' }}></div>
+        <Label name="between">
+          <CiClock1 />
+        </Label>
+        <DatePicker dateType="start"/>
+        <Label name="and">
+          <CiClock2 />
+        </Label>
+        <DatePicker dateType="end"/>
+        <Label name="interval">
+          <CiTimer />
+        </Label>
+        <SelectInterval onChange={handlePeriodIntervalChange}/>
       </$SelectorForm>
       {error !== null ? <$ErrorLabel>{error}</$ErrorLabel> : null}
     </>